+++ mail-message/data/src/main/kotlin/ch/protonmail/android/mailmessage/data/remote/resource/MessageResource.kt
@@ -69,7 +69,7 @@ data class MessageResource(
     @SerialName("NumAttachments")
     val numAttachments: Int,
     @SerialName("Flags")
-    val flags: Int,
+    val flags: Long,
     @SerialName("AttachmentInfo")
     val attachmentsInfo: AttachmentsInfoResource? = null
 ) {
